- fix latest change on SHP causing some check to be dead or optimized by the compiler - prepare more project linked files - add PR #2334 - update PR #74 #75 - sync to develop - implement internal ReShade (only work on dx renderer)

+++ src/Ext/Cell/Body.cpp
@@ -856,7 +856,7 @@ static constexpr int BridgeLevelAdd = 4;    // assembly: shl edx,2 → always 4,
 // ---------------------------------------------------------------------------
 static CellClass* ResolveCellOrWorking(int x, int y)
 {
-	return MapClass::Instance->GetCellAt(CellStruct(x, y));
+	return MapClass::Instance->GetCellAt(CellStruct((short)x, (short)y));
 }
 
 // ---------------------------------------------------------------------------
